Refactor #6 - Move index to it's own file
